batch: Move attribution unit generation
File-comparison attribution units now live in attribution_units. The
builder imports the unit model, unit ID helper, and comparison enumerator
as private collaborators.

Displayed-diff projection now imports the unit kind from the unit module
while file attribution containers remain in attribution. Batch ownership
matching stays in the builder layer.

This commit moves baseline versus working-tree unit enumeration out of
attribution and gives file-derived unit construction its own module.

The production behavior is unchanged, but the attribution builder now
focuses on metadata loading, batch alignment, and owner assignment.
