+++ src/git_stage_batch/batch/attribution.py
@@ -11,12 +11,18 @@ from __future__ import annotations
 from collections.abc import Sequence
 from contextlib import ExitStack
 from dataclasses import dataclass
-from enum import Enum
 
-from ..batch.line_mapping import LineMapping
-from ..batch.match import match_lines
+from .line_mapping import LineMapping as _LineMapping
+from .match import match_lines
 from . import attribution_fingerprints as _attribution_fingerprints
-from ..batch.query import list_batch_names, read_batch_metadata
+from .attribution_units import (
+    AttributionUnit as _AttributionUnit,
+    AttributionUnitKind as _AttributionUnitKind,
+    build_file_comparison_from_lines as _build_file_comparison_from_lines,
+    enumerate_units_from_file_comparison as _enumerate_units_from_file_comparison,
+    make_attribution_unit_id as _make_attribution_unit_id,
+)
+from .query import list_batch_names, read_batch_metadata
 from ..core.line_selection import parse_line_selection
 from ..data.consumed_selections import read_consumed_file_metadata
 from ..utils.repository_buffers import (
@@ -25,44 +31,11 @@ from ..utils.repository_buffers import (
 )
 
 
-class AttributionUnitKind(Enum):
-    """Type of attribution unit for file-centric filtering."""
-
-    PRESENCE_ONLY = "presence_only"
-    """Pure addition (no coupled deletion)."""
-
-    REPLACEMENT = "replacement"
-    """Addition coupled with deletion (modification)."""
-
-    DELETION_ONLY = "deletion_only"
-    """Pure deletion (no coupled addition)."""
-
-
-@dataclass
-class AttributionUnit:
-    """A semantic unit in the working tree file for attribution.
-
-    Represents a change unit derived from baseline ↔ working tree comparison.
-    Used for filtering diff output by determining which fragments are owned by batches.
-    """
-
-    unit_id: str
-    kind: AttributionUnitKind
-    file_path: str
-    claimed_line_in_working_tree: int | None
-    claimed_content: bytes | None
-    deletion_anchor_in_working_tree: int | None
-    deletion_content: bytes | None
-    deletion_fingerprint: _attribution_fingerprints.ContentFingerprint | None = None
-    claimed_fingerprint: _attribution_fingerprints.ContentFingerprint | None = None
-    claimed_line_count: int | None = None
-
-
 @dataclass
 class AttributedUnit:
     """An attribution unit plus the batches that currently own it."""
 
-    unit: AttributionUnit
+    unit: _AttributionUnit
     owning_batches: set[str]
 
 
@@ -75,91 +48,13 @@ class FileAttribution:
 
 
 @dataclass
-class FileComparison:
-    """Canonical baseline ↔ working-tree comparison for a file."""
-
-    file_path: str
-    baseline_lines: Sequence[bytes]
-    working_tree_lines: Sequence[bytes]
-    alignment: LineMapping
-
-    def close(self) -> None:
-        """Close the owned alignment mapping."""
-        self.alignment.close()
-
-    def __enter__(self) -> FileComparison:
-        return self
-
-    def __exit__(self, exc_type, exc, traceback) -> None:
-        self.close()
-
-
-@dataclass
 class BatchAttributionContext:
     """Reusable source alignment for one batch/file attribution pass."""
 
     batch_name: str
     file_metadata: dict
     batch_source_lines: Sequence[bytes]
-    alignment: LineMapping
-
-
-def _make_unit_id(
-    kind: AttributionUnitKind,
-    file_path: str,
-    claimed_line: int | None = None,
-    claimed_content: bytes | None = None,
-    claimed_fingerprint: _attribution_fingerprints.ContentFingerprint | None = None,
-    deletion_anchor: int | None = None,
-    deletion_content: bytes | None = None,
-    deletion_fingerprint: _attribution_fingerprints.ContentFingerprint | None = None,
-) -> str:
-    """Create a semantic, batch-independent identifier for a unit."""
-    def _hash(
-        content: bytes | None,
-        fingerprint: _attribution_fingerprints.ContentFingerprint | None = None,
-    ) -> str:
-        if fingerprint is not None:
-            if fingerprint.byte_count == 0:
-                return "none"
-            return fingerprint.sha1[:8]
-        if content is None or not content:
-            return "none"
-        return _attribution_fingerprints.fingerprint_bytes(content).sha1[:8]
-
-    if kind == AttributionUnitKind.PRESENCE_ONLY:
-        line_str = str(claimed_line) if claimed_line is not None else "missing"
-        return (
-            f"PRESENCE_ONLY:{file_path}:{line_str}:"
-            f"{_hash(claimed_content, claimed_fingerprint)}"
-        )
-
-    if kind == AttributionUnitKind.DELETION_ONLY:
-        anchor_str = str(deletion_anchor) if deletion_anchor is not None else "start"
-        return (
-            f"DELETION_ONLY:{file_path}:after-{anchor_str}:"
-            f"{_hash(deletion_content, deletion_fingerprint)}"
-        )
-
-    if kind == AttributionUnitKind.REPLACEMENT:
-        line_str = str(claimed_line) if claimed_line is not None else "missing"
-        anchor_str = str(deletion_anchor) if deletion_anchor is not None else "start"
-        return (
-            f"REPLACEMENT:{file_path}:{line_str}:after-{anchor_str}:"
-            f"c{_hash(claimed_content, claimed_fingerprint)}:"
-            f"d{_hash(deletion_content, deletion_fingerprint)}"
-        )
-
-    return f"UNKNOWN:{file_path}"
-
-
-def _single_line_content(
-    lines: Sequence[bytes],
-    line_numbers: Sequence[int],
-) -> bytes | None:
-    if len(line_numbers) != 1:
-        return None
-    return lines[line_numbers[0] - 1]
+    alignment: _LineMapping
 
 
 def _parse_presence_source_lines(file_metadata: dict) -> list[int]:
@@ -183,203 +78,6 @@ def _has_presence_source_lines(file_metadata: dict) -> bool:
         file_metadata.get("presence_claims")
         or file_metadata.get("claimed_lines")
     )
-
-
-def _build_file_comparison_from_lines(
-    file_path: str,
-    *,
-    baseline_lines: Sequence[bytes],
-    working_tree_lines: Sequence[bytes],
-) -> FileComparison:
-    alignment = match_lines(source_lines=baseline_lines, target_lines=working_tree_lines)
-    return FileComparison(
-        file_path=file_path,
-        baseline_lines=baseline_lines,
-        working_tree_lines=working_tree_lines,
-        alignment=alignment,
-    )
-
-
-def enumerate_units_from_file_comparison(
-    comparison: FileComparison,
-    units_map: dict[str, AttributionUnit],
-) -> None:
-    """Enumerate file-derived ownership units from baseline ↔ working tree comparison."""
-    file_path = comparison.file_path
-    baseline_lines = comparison.baseline_lines
-    working_tree_lines = comparison.working_tree_lines
-    alignment = comparison.alignment
-
-    baseline_to_working: dict[int, int] = {}
-    working_to_baseline: dict[int, int] = {}
-
-    for baseline_idx in range(len(baseline_lines)):
-        baseline_line_num = baseline_idx + 1
-        working_tree_line_num = alignment.get_target_line_from_source_line(baseline_line_num)
-        if working_tree_line_num is not None:
-            baseline_to_working[baseline_line_num] = working_tree_line_num
-            working_to_baseline[working_tree_line_num] = baseline_line_num
-
-    baseline_unmatched = [
-        line_num
-        for line_num in range(1, len(baseline_lines) + 1)
-        if line_num not in baseline_to_working
-    ]
-    working_unmatched = [
-        line_num
-        for line_num in range(1, len(working_tree_lines) + 1)
-        if line_num not in working_to_baseline
-    ]
-
-    baseline_runs = _group_consecutive(baseline_unmatched)
-    working_runs = _group_consecutive(working_unmatched)
-
-    paired_deletion_runs: set[int] = set()
-    paired_addition_runs: set[int] = set()
-    replacements: list[tuple[list[int], list[int], int | None]] = []
-
-    for del_run_idx, del_run in enumerate(baseline_runs):
-        if del_run_idx in paired_deletion_runs:
-            continue
-
-        del_anchor_baseline = _find_structural_predecessor(del_run, baseline_to_working)
-        del_anchor_working = None if del_anchor_baseline is None else baseline_to_working.get(del_anchor_baseline)
-
-        best_match_idx = None
-        for add_run_idx, add_run in enumerate(working_runs):
-            if add_run_idx in paired_addition_runs:
-                continue
-
-            add_anchor = _find_structural_predecessor(add_run, working_to_baseline)
-            if _anchors_match(del_anchor_working, add_anchor):
-                best_match_idx = add_run_idx
-                break
-
-        if best_match_idx is None:
-            continue
-
-        paired_deletion_runs.add(del_run_idx)
-        paired_addition_runs.add(best_match_idx)
-        replacements.append((del_run, working_runs[best_match_idx], del_anchor_working))
-
-    remaining_deletions = [
-        run for idx, run in enumerate(baseline_runs) if idx not in paired_deletion_runs
-    ]
-    remaining_additions = [
-        run for idx, run in enumerate(working_runs) if idx not in paired_addition_runs
-    ]
-
-    for del_run, add_run, anchor in replacements:
-        deletion_fingerprint = _attribution_fingerprints.fingerprint_numbered_lines(
-            baseline_lines,
-            del_run,
-        )
-        addition_content = _single_line_content(working_tree_lines, add_run)
-        addition_fingerprint = None
-        if addition_content is None:
-            addition_fingerprint = _attribution_fingerprints.fingerprint_numbered_lines(
-                working_tree_lines,
-                add_run,
-            )
-        working_tree_line = add_run[0]
-
-        unit = AttributionUnit(
-            unit_id=_make_unit_id(
-                AttributionUnitKind.REPLACEMENT,
-                file_path,
-                claimed_line=working_tree_line,
-                claimed_content=addition_content,
-                claimed_fingerprint=addition_fingerprint,
-                deletion_anchor=anchor,
-                deletion_fingerprint=deletion_fingerprint,
-            ),
-            kind=AttributionUnitKind.REPLACEMENT,
-            file_path=file_path,
-            claimed_line_in_working_tree=working_tree_line,
-            claimed_content=addition_content,
-            deletion_anchor_in_working_tree=anchor,
-            deletion_content=None,
-            deletion_fingerprint=deletion_fingerprint,
-            claimed_fingerprint=addition_fingerprint,
-            claimed_line_count=len(add_run),
-        )
-        units_map.setdefault(unit.unit_id, unit)
-
-    for del_run in remaining_deletions:
-        deletion_fingerprint = _attribution_fingerprints.fingerprint_numbered_lines(
-            baseline_lines,
-            del_run,
-        )
-        anchor_baseline = _find_structural_predecessor(del_run, baseline_to_working)
-        anchor_working = None if anchor_baseline is None else baseline_to_working.get(anchor_baseline)
-
-        unit = AttributionUnit(
-            unit_id=_make_unit_id(
-                AttributionUnitKind.DELETION_ONLY,
-                file_path,
-                deletion_anchor=anchor_working,
-                deletion_fingerprint=deletion_fingerprint,
-            ),
-            kind=AttributionUnitKind.DELETION_ONLY,
-            file_path=file_path,
-            claimed_line_in_working_tree=None,
-            claimed_content=None,
-            deletion_anchor_in_working_tree=anchor_working,
-            deletion_content=None,
-            deletion_fingerprint=deletion_fingerprint,
-        )
-        units_map.setdefault(unit.unit_id, unit)
-
-    for add_run in remaining_additions:
-        for line_num in add_run:
-            addition_content = working_tree_lines[line_num - 1]
-            unit = AttributionUnit(
-                unit_id=_make_unit_id(
-                    AttributionUnitKind.PRESENCE_ONLY,
-                    file_path,
-                    claimed_line=line_num,
-                    claimed_content=addition_content,
-                ),
-                kind=AttributionUnitKind.PRESENCE_ONLY,
-                file_path=file_path,
-                claimed_line_in_working_tree=line_num,
-                claimed_content=addition_content,
-                deletion_anchor_in_working_tree=None,
-                deletion_content=None,
-            )
-            units_map.setdefault(unit.unit_id, unit)
-
-
-def _find_structural_predecessor(
-    run: list[int],
-    line_mapping: dict[int, int],
-) -> int | None:
-    """Return the last matched line before a run, or None at start-of-file."""
-    if not run:
-        return None
-
-    for candidate in range(run[0] - 1, 0, -1):
-        if candidate in line_mapping:
-            return candidate
-    return None
-
-
-def _anchors_match(anchor1: int | None, anchor2: int | None) -> bool:
-    return anchor1 == anchor2
-
-
-def _group_consecutive(line_numbers: list[int]) -> list[list[int]]:
-    if not line_numbers:
-        return []
-
-    sorted_lines = sorted(line_numbers)
-    runs = [[sorted_lines[0]]]
-    for line in sorted_lines[1:]:
-        if line == runs[-1][-1] + 1:
-            runs[-1].append(line)
-        else:
-            runs.append([line])
-    return runs
 
 
 def build_file_attribution(file_path: str) -> FileAttribution:
@@ -409,13 +107,13 @@ def build_file_attribution(file_path: str) -> FileAttribution:
                 stack=batch_context_stack,
             )
 
-            all_units_map: dict[str, AttributionUnit] = {}
+            all_units_map: dict[str, _AttributionUnit] = {}
             with _build_file_comparison_from_lines(
                 file_path,
                 baseline_lines=baseline_lines,
                 working_tree_lines=working_tree_lines,
             ) as comparison:
-                enumerate_units_from_file_comparison(comparison, all_units_map)
+                _enumerate_units_from_file_comparison(comparison, all_units_map)
 
             if batch_contexts:
                 _enumerate_units_from_batches(
@@ -472,7 +170,7 @@ def _open_batch_attribution_contexts(
 def _enumerate_units_from_batches(
     file_path: str,
     batch_contexts: Sequence[BatchAttributionContext],
-    units_map: dict[str, AttributionUnit],
+    units_map: dict[str, _AttributionUnit],
 ) -> None:
     """Add batch-owned units that may not be visible in the working tree."""
     for batch_context in batch_contexts:
@@ -489,14 +187,14 @@ def _enumerate_units_from_batches(
 
             claimed_content = batch_source_lines[source_line - 1]
             working_tree_line = alignment.get_target_line_from_source_line(source_line)
-            unit = AttributionUnit(
-                unit_id=_make_unit_id(
-                    AttributionUnitKind.PRESENCE_ONLY,
+            unit = _AttributionUnit(
+                unit_id=_make_attribution_unit_id(
+                    _AttributionUnitKind.PRESENCE_ONLY,
                     file_path,
                     claimed_line=working_tree_line,
                     claimed_content=claimed_content,
                 ),
-                kind=AttributionUnitKind.PRESENCE_ONLY,
+                kind=_AttributionUnitKind.PRESENCE_ONLY,
                 file_path=file_path,
                 claimed_line_in_working_tree=working_tree_line,
                 claimed_content=claimed_content,
@@ -523,14 +221,14 @@ def _enumerate_units_from_batches(
                 else alignment.get_target_line_from_source_line(after_source_line)
             )
 
-            unit = AttributionUnit(
-                unit_id=_make_unit_id(
-                    AttributionUnitKind.DELETION_ONLY,
+            unit = _AttributionUnit(
+                unit_id=_make_attribution_unit_id(
+                    _AttributionUnitKind.DELETION_ONLY,
                     file_path,
                     deletion_anchor=deletion_anchor,
                     deletion_fingerprint=deletion_fingerprint,
                 ),
-                kind=AttributionUnitKind.DELETION_ONLY,
+                kind=_AttributionUnitKind.DELETION_ONLY,
                 file_path=file_path,
                 claimed_line_in_working_tree=None,
                 claimed_content=None,
@@ -542,7 +240,7 @@ def _enumerate_units_from_batches(
 
 
 def _find_owning_batches(
-    unit: AttributionUnit,
+    unit: _AttributionUnit,
     batch_contexts: Sequence[BatchAttributionContext],
 ) -> set[str]:
     """Determine which batches own a given unit."""
@@ -553,7 +251,7 @@ def _find_owning_batches(
         alignment = batch_context.alignment
         batch_source_lines = batch_context.batch_source_lines
 
-        if unit.kind == AttributionUnitKind.PRESENCE_ONLY:
+        if unit.kind == _AttributionUnitKind.PRESENCE_ONLY:
             if _batch_owns_presence_unit(
                 unit,
                 file_metadata,
@@ -561,10 +259,10 @@ def _find_owning_batches(
                 batch_source_lines,
             ):
                 owning_batches.add(batch_context.batch_name)
-        elif unit.kind == AttributionUnitKind.DELETION_ONLY:
+        elif unit.kind == _AttributionUnitKind.DELETION_ONLY:
             if _batch_owns_deletion_unit(unit, file_metadata, alignment):
                 owning_batches.add(batch_context.batch_name)
-        elif unit.kind == AttributionUnitKind.REPLACEMENT:
+        elif unit.kind == _AttributionUnitKind.REPLACEMENT:
             if (
                 _batch_owns_presence_unit(
                     unit,
@@ -580,7 +278,7 @@ def _find_owning_batches(
 
 
 def _batch_owns_presence_unit(
-    unit: AttributionUnit,
+    unit: _AttributionUnit,
     file_metadata: dict,
     alignment,
     batch_source_lines: Sequence[bytes],
@@ -622,7 +320,10 @@ def _batch_owns_presence_unit(
         )
         if source_line_range.stop - 1 > len(batch_source_lines):
             return False
-        if any(source_line not in claimed_source_line_set for source_line in source_line_range):
+        if any(
+            source_line not in claimed_source_line_set
+            for source_line in source_line_range
+        ):
             return False
         return (
             _attribution_fingerprints.fingerprint_numbered_lines(
@@ -667,8 +368,9 @@ def _batch_owns_presence_unit(
 
     return False
 
+
 def _batch_owns_deletion_unit(
-    unit: AttributionUnit,
+    unit: _AttributionUnit,
     file_metadata: dict,
     alignment,
 ) -> bool:
